src/requirements.txt
python-telegram-bot
httpx
pydantic
